refactor: simplify message part handler
